Tests/LibWeb/Text/expected/UIEvents/KeyEvent-keypress.txt

keydown A
keypress A
keydown Shift
keydown B
keypress B
